Create README.md
In mobile design, it’s important that we respect a user’s task and mindset, as well as the device’s limitations. Here you’ll learn about the general principles that can help you get started with your design.When the user interacts with their device for brief but frenzied periods of activity. Think about how you can accommodate users who are in a hurry. Make frequent, recurring tasks as efficient as possible.

Signed-off-by: SINDILE LANCELOTT TOTOBA <54500302+16lancelott@users.noreply.github.com>